# apps/api/main.py
"""FastAPI reviewer and dashboard API for the offline reference runtime."""

import logging
import os
import re
import uuid
from datetime import UTC, datetime
from pathlib import Path
from typing import Annotated

from fastapi import Depends, FastAPI, File, Form, Query, Request, UploadFile, status
from fastapi.exceptions import RequestValidationError
from fastapi.middleware.cors import CORSMiddleware
from fastapi.responses import JSONResponse

from apps.api.copilot import router as copilot_router
from apps.api.dependencies import LocalRuntime
from apps.api.schemas import (
    MAX_UPLOAD_BYTES,
    CitySummaryResponse,
    DatasetResponse,
    DecisionRequest,
    DistrictCompareRequest,
    DistrictForecastResponse,
    DistrictOverviewResponse,
    DistrictProfileResponse,
    ErrorBody,
    ErrorEnvelope,
    JobStatusResponse,
    LineageResponse,
    MappingSamplePreview,
    MappingSampleValue,
    QualityResponse,
    UploadResponse,
)
from apps.api.security import WRITE_TOKEN_HEADER, require_write_token
from apps.api.uploads import (
    aws_workflow_configured,
    get_aws_job_reference,
    get_aws_mapping_analysis,
    resume_aws_job,
)
from apps.api.uploads import (
    router as uploads_router,
)
from youth_compass import __version__
from youth_compass.domain import (
    AnalyticsNotAvailableError,
    DatasetNotFoundError,
    ForecastNotAvailableError,
    ModelInvocationError,
    QueryExecutionError,
    QueryNotPermittedError,
    SourceAcquisitionError,
    SourceNormalizationError,
    WorkflowNotFoundError,
    WorkflowStateError,
    YouthCompassError,
)
from youth_compass.domain.contracts import MappingAnalysis
from youth_compass.ports import ApprovalDecision, ForecastRequest, JobReference
from youth_compass.transformation.values import RowTransformationError, preview_column_value

_LOGGER = logging.getLogger(__name__)

_DATA_ROOT_ENV_VAR = "YOUTH_COMPASS_DATA_ROOT"
_ALLOWED_SOURCE_IPS_ENV_VAR = "YOUTH_COMPASS_ALLOWED_SOURCE_IPS"


def default_data_root() -> Path:
    """The data root to use when a caller does not name one.

    Reads the environment so a deployment can relocate the tree without
    touching code. Lambda in particular must point this at a writable location,
    because the unpacked package at /var/task is read-only.
    """
    return Path(os.environ.get(_DATA_ROOT_ENV_VAR, "data"))


def create_app(data_root: Path | None = None) -> FastAPI:
    app = FastAPI(title="New Taipei Youth Policy API", version=__version__)
    runtime = LocalRuntime(default_data_root() if data_root is None else data_root)
    app.state.runtime = runtime
    # Exposed separately so request-scoped guards can read settings without
    # reaching through the runtime.
    app.state.settings = runtime.settings

    # Lambda Function URLs expose the caller's address as the ASGI client. The
    # allowlist is opt-in so local development and the offline test suite stay
    # accessible, while the deployed stack fails closed for every other IP.
    allowed_source_ips = {
        value.strip()
        for value in os.environ.get(_ALLOWED_SOURCE_IPS_ENV_VAR, "").split(",")
        if value.strip()
    }
    if allowed_source_ips:

        @app.middleware("http")
        async def restrict_source_ip(request: Request, call_next):  # type: ignore[no-untyped-def]
            source_ip = request.client.host if request.client is not None else ""
            if source_ip not in allowed_source_ips:
                return JSONResponse(
                    status_code=status.HTTP_403_FORBIDDEN,
                    content={"detail": "source IP is not allowed"},
                )
            return await call_next(request)

    # A browser-based frontend served from another origin cannot call this API
    # unless its origin is allowed here. Defaults to empty, so nothing is
    # exposed cross-origin until a deployment configures it.
    allowed_origins = list(runtime.settings.api.allowed_origins)
    if allowed_origins:
        app.add_middleware(
            CORSMiddleware,
            allow_origins=allowed_origins,
            allow_methods=["GET", "POST", "OPTIONS"],
            # The write guard travels in a custom header, so it must be
            # allowed through preflight.
            allow_headers=["Content-Type", WRITE_TOKEN_HEADER],
            max_age=600,
        )

    app.include_router(copilot_router)
    app.include_router(uploads_router)

    @app.exception_handler(YouthCompassError)
    async def domain_error_handler(request: Request, exc: YouthCompassError) -> JSONResponse:
        del request
        status_code = _status_for_error(exc)
        trace_id = f"trc_{uuid.uuid4().hex}"
        if status_code >= status.HTTP_500_INTERNAL_SERVER_ERROR:
            # The client sees only a generic message, so the detail has to live
            # somewhere an operator can find it by trace ID.
            _LOGGER.exception("%s failed: %s", trace_id, exc)
        envelope = ErrorEnvelope(
            error=ErrorBody(
                code=_error_code(exc),
                message=_client_message(exc, status_code),
                trace_id=trace_id,
            )
        )
        return JSONResponse(
            status_code=status_code,
            content=envelope.model_dump(mode="json", by_alias=True),
        )

    @app.exception_handler(RequestValidationError)
    async def validation_error_handler(
        request: Request, exc: RequestValidationError
    ) -> JSONResponse:
        del request
        details: list[dict[str, object]] = [
            {
                "field": ".".join(str(part) for part in error["loc"]),
                "reason": error["msg"],
            }
            for error in exc.errors()
        ]
        envelope = ErrorEnvelope(
            error=ErrorBody(
                code="VALIDATION_FAILED",
                message="request validation failed",
                details=details,
                trace_id=f"trc_{uuid.uuid4().hex}",
            )
        )
        return JSONResponse(
            status_code=status.HTTP_422_UNPROCESSABLE_CONTENT,
            content=envelope.model_dump(mode="json", by_alias=True),
        )

    @app.get("/health", tags=["operations"])
    def health() -> dict[str, str]:
        return {"status": "ok", "version": __version__}

    @app.post(
        "/api/v1/datasets/upload",
        response_model=UploadResponse,
        status_code=status.HTTP_202_ACCEPTED,
        tags=["reviewer"],
        dependencies=[Depends(require_write_token)],
    )
    async def upload_dataset(
        request: Request,
        file: Annotated[UploadFile, File()],
        submitted_by: Annotated[str, Form()],
        topic_hint: Annotated[str | None, Form()] = None,
    ) -> UploadResponse:
        content = await file.read(MAX_UPLOAD_BYTES + 1)
        if len(content) > MAX_UPLOAD_BYTES:
            raise WorkflowStateError("upload exceeds the 25 MiB local limit")
        reference = _runtime(request).workflow.submit_bytes(
            file_name=file.filename or "",
            content=content,
            submitted_by=submitted_by,
            topic_hint=topic_hint,
        )
        return UploadResponse(
            job_id=reference.job_id,
            status=reference.status,
            links={"job": f"/api/v1/ingestion-jobs/{reference.job_id}"},
        )

    @app.get(
        "/api/v1/ingestion-jobs/{job_id}",
        response_model=JobStatusResponse,
        tags=["reviewer"],
    )
    def get_ingestion_job(job_id: str, request: Request) -> JobStatusResponse:
        try:
            return _job_response(_runtime(request), job_id)
        except WorkflowNotFoundError:
            if not aws_workflow_configured():
                raise
            return _aws_job_response(get_aws_job_reference(request, job_id))

    @app.get(
        "/api/v1/ingestion-jobs/{job_id}/mapping",
        response_model=MappingAnalysis,
        tags=["reviewer"],
    )
    def get_mapping(job_id: str, request: Request) -> MappingAnalysis:
        return _mapping_for_job(request, job_id)

    @app.get(
        "/api/v1/ingestion-jobs/{job_id}/mapping-preview",
        response_model=list[MappingSamplePreview],
        tags=["reviewer"],
    )
    def get_mapping_preview(job_id: str, request: Request) -> list[MappingSamplePreview]:
        """Show bounded before/after examples without exposing source rows or paths."""

        analysis = _mapping_for_job(request, job_id)
        profiles = {column.name: column for column in analysis.profile.columns}
        previews: list[MappingSamplePreview] = []
        for mapping in analysis.proposal.columns:
            samples: list[MappingSampleValue] = []
            profile = profiles.get(mapping.source_column)
            for source in profile.sample_values[:3] if profile is not None else []:
                try:
                    canonical = _public_preview_values(preview_column_value(source, mapping))
                except RowTransformationError as exc:
                    samples.append(MappingSampleValue(source=source, error=str(exc)[:300]))
                else:
                    samples.append(MappingSampleValue(source=source, canonical=canonical))
            previews.append(
                MappingSamplePreview(
                    source_column=mapping.source_column,
                    target_field=mapping.target_field,
                    transformation=mapping.transformation,
                    samples=samples,
                )
            )
        return previews

    @app.post(
        "/api/v1/ingestion-jobs/{job_id}/decision",
        response_model=JobStatusResponse,
        tags=["reviewer"],
        dependencies=[Depends(require_write_token)],
    )
    def decide_ingestion(
        job_id: str, payload: DecisionRequest, request: Request
    ) -> JobStatusResponse:
        runtime = _runtime(request)
        decision = ApprovalDecision(
            approved=payload.decision == "approve",
            decided_by=payload.decided_by,
            decided_at=datetime.now(UTC),
            notes=payload.comment,
        )
        try:
            runtime.workflow.resume_after_approval(job_id, decision)
            return _job_response(runtime, job_id)
        except WorkflowNotFoundError:
            if not aws_workflow_configured():
                raise
            resume_aws_job(request, job_id, decision)
            return _aws_job_response(get_aws_job_reference(request, job_id))

    @app.get(
        "/api/v1/ingestion-jobs/{job_id}/quality-report",
        response_model=QualityResponse,
        tags=["reviewer"],
    )
    def get_quality_report(job_id: str, request: Request) -> QualityResponse:
        job = _runtime(request).workflow.get_job(job_id)
        if job.manifest is None:
            raise WorkflowStateError(f"quality report is not available while job is {job.status}")
        return QualityResponse(
            job_id=job.job_id,
            dataset_id=job.dataset_id,
            dataset_version=job.manifest.dataset_version,
            publication_status=job.manifest.status.value,
            quality=job.manifest.quality,
        )

    @app.get("/api/v1/datasets", response_model=list[DatasetResponse], tags=["catalog"])
    def list_datasets(request: Request) -> list[DatasetResponse]:
        return [
            DatasetResponse.from_metadata(metadata)
            for metadata in _runtime(request).catalog.list_datasets()
        ]

    @app.get(
        "/api/v1/datasets/{dataset_id}",
        response_model=DatasetResponse,
        tags=["catalog"],
    )
    def get_dataset(dataset_id: str, request: Request) -> DatasetResponse:
        return DatasetResponse.from_metadata(_runtime(request).catalog.get(dataset_id))

    @app.get(
        "/api/v1/datasets/{dataset_id}/versions",
        response_model=list[DatasetResponse],
        tags=["catalog"],
    )
    def get_dataset_versions(dataset_id: str, request: Request) -> list[DatasetResponse]:
        versions = _runtime(request).catalog.list_versions(dataset_id)
        if not versions:
            raise DatasetNotFoundError(dataset_id)
        return [DatasetResponse.from_metadata(metadata) for metadata in versions]

    @app.get(
        "/api/v1/datasets/{dataset_id}/lineage",
        response_model=LineageResponse,
        tags=["catalog"],
    )
    def get_lineage(dataset_id: str, request: Request) -> LineageResponse:
        metadata = _runtime(request).catalog.get(dataset_id)
        return LineageResponse(
            dataset_id=metadata.dataset_id,
            dataset_version=metadata.version,
            source_sha256=metadata.source_sha256,
            mapping_version=metadata.mapping_version,
            approved_by=metadata.approved_by,
            created_at=metadata.created_at,
            published_at=metadata.published_at,
        )

    @app.get(
        "/api/v1/city/summary",
        response_model=CitySummaryResponse,
        tags=["dashboard"],
    )
    def city_summary(
        request: Request,
        dataset_id: str = Query(..., alias="datasetId"),
        metric_code: str = Query(..., alias="metricCode"),
        period: str | None = Query(default=None),
    ) -> CitySummaryResponse:
        result = _runtime(request).analytics(dataset_id).city_summary(metric_code, period)
        return CitySummaryResponse.from_result(result)

    @app.get(
        "/api/v1/districts",
        response_model=DistrictProfileResponse,
        tags=["dashboard"],
    )
    def list_districts(
        request: Request,
        dataset_id: str = Query(..., alias="datasetId"),
        metric_code: str = Query(..., alias="metricCode"),
        period: str | None = Query(default=None),
    ) -> DistrictProfileResponse:
        result = (
            _runtime(request)
            .analytics(dataset_id)
            .district_profile(
                metric_code,
                period=period,
            )
        )
        return DistrictProfileResponse.from_result(result)

    @app.get(
        "/api/v1/districts/{district_code}/profile",
        response_model=DistrictProfileResponse,
        tags=["dashboard"],
    )
    def district_profile(
        district_code: str,
        request: Request,
        dataset_id: str = Query(..., alias="datasetId"),
        metric_code: str = Query(..., alias="metricCode"),
        period: str | None = Query(default=None),
    ) -> DistrictProfileResponse:
        result = (
            _runtime(request)
            .analytics(dataset_id)
            .district_profile(
                metric_code,
                district_codes=[district_code],
                period=period,
            )
        )
        return DistrictProfileResponse.from_result(result)

    @app.get(
        "/api/v1/districts/{district_code}/overview",
        response_model=DistrictOverviewResponse,
        tags=["dashboard"],
    )
    def district_overview(
        district_code: str,
        request: Request,
        dataset_id: str = Query(default="population", alias="datasetId"),
        metric_code: str = Query(default="population_count", alias="metricCode"),
    ) -> DistrictOverviewResponse:
        result = (
            _runtime(request)
            .analytics(dataset_id)
            .district_overview(district_code, metric_code=metric_code)
        )
        return DistrictOverviewResponse.from_result(result)

    @app.get(
        "/api/v1/districts/{district_code}/forecast",
        response_model=DistrictForecastResponse,
        tags=["dashboard"],
    )
    def district_forecast(
        district_code: str,
        request: Request,
        metric_code: str = Query(default="population_count", alias="metricCode"),
        horizon_years: int = Query(default=5, alias="horizonYears", ge=1, le=10),
    ) -> DistrictForecastResponse:
        service = _runtime(request).forecasts()
        if service is None:
            raise ForecastNotAvailableError("no forecast provider is configured")
        result = service.get_forecast(
            ForecastRequest(
                metric_code=metric_code,
                district_codes=[district_code],
                horizon_years=horizon_years,
            )
        )
        return DistrictForecastResponse.from_result(district_code, result)

    @app.post(
        "/api/v1/districts/compare",
        response_model=DistrictProfileResponse,
        tags=["dashboard"],
    )
    def compare_districts(
        payload: DistrictCompareRequest, request: Request
    ) -> DistrictProfileResponse:
        result = (
            _runtime(request)
            .analytics(payload.dataset_id)
            .district_profile(
                payload.metric_code,
                district_codes=payload.district_codes,
                period=payload.period,
            )
        )
        return DistrictProfileResponse.from_result(result)

    return app


def _runtime(request: Request) -> LocalRuntime:
    runtime: LocalRuntime = request.app.state.runtime
    return runtime


def _mapping_for_job(request: Request, job_id: str) -> MappingAnalysis:
    """Resolve local or AWS-backed review state through one public endpoint."""

    try:
        analysis = _runtime(request).workflow.get_job(job_id).mapping_analysis
    except WorkflowNotFoundError:
        if not aws_workflow_configured():
            raise
        analysis = get_aws_mapping_analysis(request, job_id)

    # Profiling needs the real source path internally, but reviewer APIs must not
    # disclose local/Lambda storage layout. Preserve the canonical contract while
    # reducing the public value to the user-visible file name.
    public_profile = analysis.profile.model_copy(
        update={"source_path": Path(analysis.profile.file_name)}
    )
    return analysis.model_copy(update={"profile": public_profile})


def _public_preview_values(
    values: dict[str, object],
) -> dict[str, str | int | float | bool | None]:
    """Constrain transformation output to the JSON scalar preview contract."""

    result: dict[str, str | int | float | bool | None] = {}
    for key, value in values.items():
        if value is None or isinstance(value, (str, int, float, bool)):
            result[key] = value
        else:
            result[key] = str(value)
    return result


def _job_response(runtime: LocalRuntime, job_id: str) -> JobStatusResponse:
    job = runtime.workflow.get_job(job_id)
    issues = [issue.message for issue in job.mapping_analysis.validation.issues]
    quality_score = (
        job.manifest.quality.quality_score if job.manifest else job.metadata.quality_score
    )
    return JobStatusResponse(
        job_id=job.job_id,
        dataset_id=job.dataset_id,
        status=job.status,
        source_format=job.source_format,
        current_step=job.status.value,
        quality_score=quality_score,
        created_at=job.metadata.created_at,
        warnings=[*job.source_warnings, *job.mapping_analysis.proposal.warnings, *issues],
        links={
            "mapping": f"/api/v1/ingestion-jobs/{job.job_id}/mapping",
            "quality": f"/api/v1/ingestion-jobs/{job.job_id}/quality-report",
        },
    )


def _aws_job_response(reference: JobReference) -> JobStatusResponse:
    return JobStatusResponse(
        job_id=reference.job_id,
        status=reference.status,
        current_step=reference.status.value,
        created_at=reference.created_at,
        links={"job": f"/api/v1/ingestion-jobs/{reference.job_id}"},
    )


def _status_for_error(exc: YouthCompassError) -> int:
    if isinstance(
        exc,
        DatasetNotFoundError
        | AnalyticsNotAvailableError
        | WorkflowNotFoundError
        | ForecastNotAvailableError,
    ):
        return status.HTTP_404_NOT_FOUND
    if isinstance(exc, WorkflowStateError | QueryNotPermittedError):
        return status.HTTP_409_CONFLICT
    if isinstance(exc, QueryExecutionError):
        return status.HTTP_422_UNPROCESSABLE_CONTENT
    if isinstance(exc, ModelInvocationError):
        return status.HTTP_503_SERVICE_UNAVAILABLE
    if isinstance(exc, SourceNormalizationError | SourceAcquisitionError):
        return status.HTTP_422_UNPROCESSABLE_CONTENT
    return status.HTTP_500_INTERNAL_SERVER_ERROR


#: Server-side detail that must not travel to a client: absolute filesystem
#: paths, S3 URIs, and ARNs all appear verbatim inside adapter error text.
_INTERNAL_DETAIL = re.compile(
    r"arn:[a-z0-9-]*:[^\s,;'\"]+"
    r"|s3://[^\s,;'\"]+"
    r"|(?<![\w.])/(?:[\w.+-]+/)+[\w.+-]*"
)


def _client_message(exc: YouthCompassError, status_code: int) -> str:
    """Describe a failure without handing the caller internal topology.

    A 4xx tells the caller what they must change, so its text is kept and only
    redacted. A 5xx is our fault, and its text carries adapter detail (paths,
    ARNs, AWS access-denied specifics) the caller can neither act on nor is
    entitled to, so it collapses to a fixed message; the trace ID and the server
    logs remain the way to correlate it.
    """

    if status_code >= status.HTTP_500_INTERNAL_SERVER_ERROR:
        return "the request could not be completed"
    return _INTERNAL_DETAIL.sub("[redacted]", str(exc))


def _error_code(exc: YouthCompassError) -> str:
    name = type(exc).__name__.removesuffix("Error")
    snake = "".join(f"_{character}" if character.isupper() else character for character in name)
    return snake.upper().lstrip("_")


app = create_app()