# contracts/api/openapi.json
{
  "components": {
    "schemas": {
...
      "CitySummaryResponse": {
        "properties": {
          "datasetId": {
            "title": "Datasetid",
            "type": "string"
          },
          "datasetVersion": {
            "title": "Datasetversion",
            "type": "string"
          },
          "districtCount": {
            "title": "Districtcount",
            "type": "integer"
          },
          "estimatedValue": {
            "title": "Estimatedvalue",
            "type": "number"
          },
          "metricCode": {
            "title": "Metriccode",
            "type": "string"
          },
          "period": {
            "title": "Period",
            "type": "string"
          },
          "populationScope": {
            "title": "Populationscope",
            "type": "string"
          },
          "qualityScore": {
            "title": "Qualityscore",
            "type": "number"
          },
          "unitCode": {
            "title": "Unitcode",
            "type": "string"
          },
          "value": {
            "title": "Value",
            "type": "number"
          }
        },
        "required": [
          "datasetId",
          "datasetVersion",
          "metricCode",
          "period",
          "value",
          "unitCode",
          "populationScope",
          "districtCount",
          "estimatedValue",
          "qualityScore"
        ],
        "title": "CitySummaryResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "New Taipei Youth Policy API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/city/summary": {
      "get": {
        "operationId": "city_summary_api_v1_city_summary_get",
        "parameters": [
          {
            "in": "query",
            "name": "datasetId",
            "required": true,
            "schema": {
              "title": "Datasetid",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "metricCode",
            "required": true,
            "schema": {
              "title": "Metriccode",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "period",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Period"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CitySummaryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "City Summary",
        "tags": [
          "dashboard"
        ]
      }
    },
...
    "/api/v1/datasets/upload": {
      "post": {
        "operationId": "upload_dataset_api_v1_datasets_upload_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_dataset_api_v1_datasets_upload_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/apps__api__schemas__UploadResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Upload Dataset",
        "tags": [
          "reviewer"
        ]
      }
    },
    "/api/v1/datasets/{dataset_id}": {
      "get": {
        "operationId": "get_dataset_api_v1_datasets__dataset_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "dataset_id",
            "required": true,
            "schema": {
              "title": "Dataset Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatasetResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Dataset",
        "tags": [
          "catalog"
        ]
      }
    },
    "/api/v1/datasets/{dataset_id}/lineage": {
      "get": {
        "operationId": "get_lineage_api_v1_datasets__dataset_id__lineage_get",
        "parameters": [
          {
            "in": "path",
            "name": "dataset_id",
            "required": true,
            "schema": {
              "title": "Dataset Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LineageResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Lineage",
        "tags": [
          "catalog"
        ]
      }
    },
    "/api/v1/datasets/{dataset_id}/versions": {
      "get": {
        "operationId": "get_dataset_versions_api_v1_datasets__dataset_id__versions_get",
        "parameters": [
          {
            "in": "path",
            "name": "dataset_id",
            "required": true,
            "schema": {
              "title": "Dataset Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DatasetResponse"
                  },
                  "title": "Response Get Dataset Versions Api V1 Datasets  Dataset Id  Versions Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Dataset Versions",
        "tags": [
          "catalog"
        ]
      }
    },
    "/api/v1/districts": {
      "get": {
        "operationId": "list_districts_api_v1_districts_get",
        "parameters": [
          {
            "in": "query",
            "name": "datasetId",
            "required": true,
            "schema": {
              "title": "Datasetid",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "metricCode",
            "required": true,
            "schema": {
              "title": "Metriccode",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "period",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Period"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DistrictProfileResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Districts",
        "tags": [
          "dashboard"
        ]
      }
    },
    "/api/v1/districts/compare": {
      "post": {
        "operationId": "compare_districts_api_v1_districts_compare_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DistrictCompareRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DistrictProfileResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Compare Districts",
        "tags": [
          "dashboard"
        ]
      }
    },
    "/api/v1/districts/{district_code}/forecast": {
      "get": {
        "operationId": "district_forecast_api_v1_districts__district_code__forecast_get",
        "parameters": [
          {
            "in": "path",
            "name": "district_code",
            "required": true,
            "schema": {
              "title": "District Code",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "metricCode",
            "required": false,
            "schema": {
              "default": "population_count",
              "title": "Metriccode",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "horizonYears",
            "required": false,
            "schema": {
              "default": 5,
              "maximum": 10,
              "minimum": 1,
              "title": "Horizonyears",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DistrictForecastResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "District Forecast",
        "tags": [
          "dashboard"
        ]
      }
    },
    "/api/v1/districts/{district_code}/overview": {
      "get": {
        "operationId": "district_overview_api_v1_districts__district_code__overview_get",
        "parameters": [
          {
            "in": "path",
            "name": "district_code",
            "required": true,
            "schema": {
              "title": "District Code",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "datasetId",
            "required": false,
            "schema": {
              "default": "population",
              "title": "Datasetid",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "metricCode",
            "required": false,
            "schema": {
              "default": "population_count",
              "title": "Metriccode",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DistrictOverviewResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "District Overview",
        "tags": [
          "dashboard"
        ]
      }
    },
...
  }
}